fetch the own votes and the number votes via ajax
for each bingo_field.id, there is a list [vote, num_votes] in the
ajax response, which contains:
* the own vote ("+", "0", "-")
* the total number of votes for this word in the game.

if there is a bingo_board parameter, the field ids for the given board
will be fetched, so its possible to view the boards of other users.

For inactive games, the update script is not included in the template,
because they do not change anyway.