allow choosing rating="------" in admin interface
without blank=True, "no rating" is selected for unrated boards, but the
board cannot be saved with "no rating". It still cannot be saved with
"no rating", as the value of the choise is "None", but with "-----", it
can be set to None and "-----" is selected for unrated fields. "-----"
has value="", which works fine for setting the IntegerField to None.