fixed error on missing key in cached vote counts
the vote count cache contains all the words, which are on any board at
the time when it's created. For normal words, every word should be
associated with every bingoboard, even when not on board. But the middle
words are only associated with a board (and via the board with a game),
when they are randomly chosen. So the cache from the first board does
not contain the middle words of the next boards, which gives an KeyError

Fix at multiple places:
return 0 for all middle words, exclude middle words from cache on
creation, recreate the cache for missing word ids and use .get(index, 0)