Lowered the threshold for a word to appear in the topic list to 60%
Boards that are created but not used lower the percent of clicks per
word, so 80% often excludes relevant topics.

A long term fix would be to exclude unused boards (either with no
clicked words or boards that were not active until the end of the game)
from the wordvote calculation.